Add hidden preference to use a split column guide in commit messages
The conventional git commit message styling says to limit your first line
to roughly 50 characters, and subsequent lines to 76 characters. The column
guide in GitX doesn't support this.

Implement 2 new hidden preferences, one to determine whether it uses a split
column guide and the other to set the length of this guide. Default these
preferences to NO and 76, respectively. When the split column guide is
turned on, the drawing code will draw the normal column guide only for the
first line and it will use the split column guide length for every other
line.